+++ HISTORY.rst
@@ -1,10 +1,14 @@
 0.5.1 (2015-??-??)
 ------------------
 
-- Make it possible to pass a language tag to a 
+- Make it possible to pass a language tag to 
   :meth:`skosprovider.registry.Registry.find` that will be passed on to all
   relevant registered providers. This determines in what language the
-  labels of the returned concepts will de displayed. (#10) [dieuska]
+  labels of the returned concepts will displayed. (#10) [dieuska]
+- Make it possible to pass a language tag to 
+  :meth:`skosprovider.registry.Registry.get_all` that will be passed on to all
+  registered providers. This determines in what language the
+  labels of the returned concepts will displayed.
 
 0.5.0 (2014-12-18)
 ------------------